requirements.txt
numpy
matplotlib
torch
torchvision
Pillow
tqdm